Adding status query for the current user id
Adding support for open status queries and secure status queries. A first open status query that's required is the query for the current user id. This query is accessible regardless of the login state. If the user is not logged in, the query returns -1. If the user is logged int, the query returns the user id of the current user.