Adding explicit inclusion of log4j-slf4j-impl to have the latest version of spring boot at least log errors with jdk11
Since the update to jdk11 and the resulting update to the latest version of spring boot, the app started crashing in quite a few places. The latest annoyance is the lack of logging. This is a workaround that makes it log errors. I'm still wondering how to make it log INFO - it doesn't seem to be reading the settings in the applocation.properties file any longer.